requirements.txt
flask==2.0.1
pycryptodome==3.10.1
cryptography==3.4.7 